Draft implementation of the PerfectHash function and small fixed in the Universal Hash function.
This commint introduce a couple of improvemnt to the hash functions, in particular, some improvment are about Universal Hash function and other on the PerferctHash.

In conclusion, some bugs or "alert" arou found and it is better fixed it in the next work.

the bugs funds are that the class not pointert (no auto elem = new *) can not assign, or better can have some error with the assign operator.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>